pyproject.toml
[tool.poetry]
name = "test-task-drf"
version = "0.1.0"
description = "Test task Django REST framework "
authors = ["Nikita <nvk.mpei@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-drf = "^1.1.3"
factory-boy = "^3.3.0"
pytest-django = "^4.5.2"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
extend-exclude = '''
(
    migrations
)
'''

[tool.flake8]
max-line-length = 120
exclude = [
    ".*",
    "__pycache__",
    "data/",
    "static/",
    "logs/",]

[tool.isort]
profile = "black"
line_length = 120
skip_glob = ["**/migrations/*", "**/settings/*"]
